pyproject.toml
[project]
name = "bertopic-pro"
version = "0.1.0"
description = "Professional BERTopic-based text analysis desktop application"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    # Desktop UI Framework
    "PySide6>=6.5.0",
    "PySide6-Addons>=6.5.0",

    # BERTopic & ML Dependencies
    "bertopic>=0.16.0",
    "sentence-transformers>=2.2.0",
    "umap-learn>=0.5.3",
    "hdbscan>=0.8.33",
    "scikit-learn>=1.3.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",

    # NLP & Text Processing
    "jieba>=0.42.1",
    "numpy>=1.24.0",
    "pandas>=2.0.0",

    # Visualization
    "plotly>=5.17.0",

    # Data Import/Export
    "openpyxl>=3.1.2",
    "xlsxwriter>=3.1.9",

    # LLM Integration
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",

    # Model Management
    "huggingface-hub>=0.19.0",

    # Utilities
    "tqdm>=4.66.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "chardet>=5.0.0",
    "joblib>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-qt>=4.2.0",
    "black>=23.0.0",
    "pyinstaller>=6.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
